+++ js/app/controllers/sidebar/artistdetailscontroller.js
@@ -10,13 +10,14 @@
 
 
 angular.module('Music').controller('ArtistDetailsController', [
-	'$rootScope', '$scope', 'Restangular', 'gettextCatalog', 'libraryService',
-	function ($rootScope, $scope, Restangular, gettextCatalog, libraryService) {
+	'$rootScope', '$scope', 'Restangular', 'gettextCatalog', 'libraryService', 'playlistService',
+	function ($rootScope, $scope, Restangular, gettextCatalog, libraryService, playlistService) {
 
 		function resetContents() {
 			$scope.artist = null;
+			$scope.featuredAlbums = null;
+			$scope.artistTracks = null;
 			$scope.artistAlbumTrackCount = 0;
-			$scope.artistTrackCount = 0;
 			$scope.loading = true;
 			$scope.artAvailable = false;
 			$scope.lastfmInfo = null;
@@ -36,7 +37,12 @@ angular.module('Music').controller('ArtistDetailsController', [
 
 				$scope.artist = libraryService.getArtist(artistId);
 				$scope.artistAlbumTrackCount = _($scope.artist.albums).map('tracks').flatten().size();
-				$scope.artistTrackCount = libraryService.findTracksByArtist(artistId).length;
+				$scope.artistTracks = libraryService.findTracksByArtist(artistId);
+				$scope.featuredAlbums = _($scope.artistTracks).map('album').uniq().difference($scope.artist.albums).value();
+
+				if ($scope.selectedTab == 'tracks' && $scope.artistTracks.length == 0) {
+					$scope.selectedTab = 'info';
+				}
 
 				let art = $('#app-sidebar .albumart');
 				art.css('background-image', '');
@@ -93,6 +99,62 @@ angular.module('Music').controller('ArtistDetailsController', [
 			}
 		}
 
+		$scope.getAlbumData = function(listItem, index, _scope) {
+			return {
+				title: listItem.name,
+				title2: listItem.year,
+				tooltip: listItem.name,
+				number: index + 1,
+				id: listItem.id,
+				art: listItem
+			};
+		};
+
+		$scope.getAlbumDraggable = function(albumId) {
+			return { album: albumId };
+		};
+
+		$scope.onAlbumClick = function(albumId) {
+			// TODO: play/pause if currently playing album clicked?
+			const album = libraryService.getAlbum(albumId);
+			playTracks('album-' + album.id, album.tracks);
+		};
+
+		$scope.getTrackData = function(listItem, index, _scope) {
+			return {
+				title: listItem.title,
+				title2: listItem.album.name + (listItem.album.year ? ` (${listItem.album.year})` : ''),
+				tooltip: listItem.title,
+				tooltip2: listItem.album.name + (listItem.album.year ? ` (${listItem.album.year})` : ''),
+				number: index + 1,
+				id: listItem.id,
+				art: listItem.album
+			};
+		};
+
+		$scope.getTrackDraggable = function(trackId) {
+			return { track: trackId };
+		};
+
+		$scope.onTrackClick = function(trackId, index) {
+			const currentTrack = $scope.$parent.currentTrack;
+			if (currentTrack?.id === trackId && currentTrack?.type == 'song') {
+				// play/pause if currently playing list item clicked
+				playlistService.publish('togglePlayback');
+			} else {
+				// on any other list item, start playing the list from this item
+				playTracks('artist-tracks-' + $scope.artist.id, $scope.artistTracks, index);
+			}
+		};
+
+		function playTracks(listId, tracks, startIndex /*optional*/) {
+			let playlist = _.map(tracks, function(track) {
+				return { track: track };
+			});
+			playlistService.setPlaylist(listId, playlist, startIndex);
+			playlistService.publish('play');
+		}
+
 		$scope.onShowAllSimilar = function() {
 			$scope.allSimilarShown = true;
 			$scope.allSimilarLoading = true;
@@ -106,7 +168,7 @@ angular.module('Music').controller('ArtistDetailsController', [
 		};
 
 		function setSimilarArtists(artists) {
-			// siliar artists are divided to those within the library and the rest
+			// similar artists are divided to those within the library and the rest
 			let artistIsInLib = function(artist) {
 				return 'id' in artist && artist.id !== null;
 			};
@@ -127,5 +189,7 @@ angular.module('Music').controller('ArtistDetailsController', [
 				resetContents();
 			}
 		});
+
+		$scope.$watch('selectedTab', $scope.$parent.adjustFixedPositions);
 	}
 ]);